Revert "profile screen updated and learning page is also updated"
