+++ Cybware/ProfileTableViewController.swift
@@ -1,16 +1,32 @@
 //
-//  TopicTableViewController.swift
+//  ProfileTableViewController.swift
 //  Cybware
 //
-//  Created by Shayaan Haque on 22/05/24.
+//  Created by Batch-2 on 21/05/24.
 //
 
 import UIKit
 
-class TopicTableViewController: UITableViewController {
+class ProfileTableViewController: UITableViewController {
+    
+    
+    required init?(coder: NSCoder)
+    {
+        super.init(coder: coder)
+        self.tabBarItem.title = "Profile"
+        self.tabBarItem.image = UIImage(systemName: "person")
+    }
 
+    @IBOutlet var profile_picture: UIImageView!
     override func viewDidLoad() {
         super.viewDidLoad()
+        
+        profile_picture.layer.cornerRadius =
+        profile_picture.frame.size.width / 5
+        profile_picture.clipsToBounds = true
+        profile_picture.layer.borderColor = UIColor.white.cgColor
+        profile_picture.layer.borderWidth = 3
+
 
         // Uncomment the following line to preserve selection between presentations
         // self.clearsSelectionOnViewWillAppear = false
@@ -21,15 +37,15 @@ class TopicTableViewController: UITableViewController {
 
     // MARK: - Table view data source
 
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        // #warning Incomplete implementation, return the number of sections
-        return 1
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return
-    }
+//    override func numberOfSections(in tableView: UITableView) -> Int {
+//        // #warning Incomplete implementation, return the number of sections
+//        return 1
+//    }
+//
+//    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+//        // #warning Incomplete implementation, return the number of rows
+//        return row.count
+//    }
 
     /*
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
